--- Samples/Timers/Timers.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2036
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Timers", "Timers\Timers.csproj", "{A7EB510D-E297-4E0E-9CFC-15D6E9713208}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A7EB510D-E297-4E0E-9CFC-15D6E9713208}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A7EB510D-E297-4E0E-9CFC-15D6E9713208}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A7EB510D-E297-4E0E-9CFC-15D6E9713208}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A7EB510D-E297-4E0E-9CFC-15D6E9713208}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7BDDB3E1-F6BC-4EEC-A6DE-48A84E9CA8B8}
-	EndGlobalSection
-EndGlobal
