--- Samples/Timers/Timers/Timers.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Reference Include="Microsoft.PSharp">
-      <HintPath>..\..\..\bin\net46\Microsoft.PSharp.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
-</Project>
